+++ src/test/groovy/ru/vyarus/dropwizard/guice/debug/WebMappingsReportTest.groovy
@@ -1,0 +1,39 @@
+package ru.vyarus.dropwizard.guice.debug
+
+import io.dropwizard.Application
+import io.dropwizard.Configuration
+import io.dropwizard.setup.Bootstrap
+import io.dropwizard.setup.Environment
+import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.support.web.WebModule
+import ru.vyarus.dropwizard.guice.test.spock.UseDropwizardApp
+import spock.lang.Specification
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 22.10.2019
+ */
+@UseDropwizardApp(App)
+class WebMappingsReportTest extends Specification {
+
+    def "Check web report"() {
+        // actual reporting checked manually (test used for reporting configuration)
+
+        expect: "check report doesn't fail"
+        true
+    }
+
+    static class App extends Application<Configuration> {
+        @Override
+        void initialize(Bootstrap<Configuration> bootstrap) {
+            bootstrap.addBundle(GuiceBundle.builder()
+                    .modules(new WebModule())
+                    .printWebMappings()
+                    .build())
+        }
+
+        @Override
+        void run(Configuration configuration, Environment environment) throws Exception {
+        }
+    }
+}