+++ CHANGELOG.md
@@ -144,7 +144,8 @@
     - Add guice bindings report (`printGuiceBindings()` or `printAllGuiceBindings()`)
     - Add guice aop appliance report (`.printGuiceAopMap()`). This report supposed to be used as "a tool" to look exact 
       services and so configurable method version is directly available: 
-      `.printGuiceAopMap(new GuiceAopConfig().types(...).methods(...))`      
+      `.printGuiceAopMap(new GuiceAopConfig().types(...).methods(...))`
+    - Add web mappings report (`.printWebMappings()`): prints all registered servlets and filters        
 * Fix configuration bindings for recursive configuration object declarations (#60)
 * Guicey version added into BOM (dependencyManagement section in guicey pom) to avoid duplicate versions declarations
 * Java 11 compatibility. Automatic module name (in meta-inf): `dropwizard-guicey.core`